Remove ParamSpecBase notion from sweep
Recently ParamSpecBase class was moved in qcodes to a different file/module.
This was not accounted for here. Hence a user had an import problem because
ParamSpecBase in newer qcodes (0.4.0) is in a different module that it was before.

Then, because ParamSpecBase is actually not used at all by sweep, this commit
also completely removes it, and with that the dependency on ParamSpecBase
is completely gone.